refactor: add AssistantClient and move AssistantRuntimeProvider to use AssistantClient
