+++ packages/react/src/client/AssistantToolUIClient.ts
@@ -1,0 +1,35 @@
+import { resource, tapState } from "@assistant-ui/tap";
+import { tapApi } from "../utils/tap-store";
+import {
+  AssistantToolUIState,
+  AssistantToolUIActions,
+} from "./types/AssistantToolUI";
+
+export const AssistantToolUIClient = resource(() => {
+  const [state, setState] = tapState<AssistantToolUIState>(() => ({}));
+
+  const api = tapApi<AssistantToolUIState, AssistantToolUIActions>(state, {
+    setToolUI: (toolName, render) => {
+      setState((prev) => {
+        return {
+          ...prev,
+          [toolName]: [...(prev[toolName] ?? []), render],
+        };
+      });
+
+      return () => {
+        setState((prev) => {
+          return {
+            ...prev,
+            [toolName]: prev[toolName]?.filter((r) => r !== render) ?? [],
+          };
+        });
+      };
+    },
+  });
+
+  return {
+    state,
+    api,
+  };
+});
